Add supervisor package into mcav_autonomy (#18)
* Add supervisor package into mcav_autonomy

The purpose of this package is to monitor the health of sensors and nodes. This is being merged into the main mcav_autonomy repository to keep it closer to the nodes and topics that it monitors. It will make it easy to work with in preparation for the addition of new features.

This closes issue #11

* Remove installation and usage documentation

---------

Co-authored-by: Srihari Sridhar <ssrihari2002@gmail.com>